Update soup.py
Starting to get information. We can pull the data from alpha query with this, the last part of the tables need to be stripped of html, but it does work. In order to work with yahoo finance I have to use selenium webdriver, I used edge as the driver option so we might have to change it for others.